+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Any widget can be added to a progress layout, not just the built-in cell types
   - TODO New cell: timeElapsed
   - Added `compact` style to `timeRemaining` cells.
+- Added `Crop` widget that can crop or pad another widget to a fixed size
 
 ### Changed
 - When setting conflicting styles on a `Table` or its cells, the innermost style now takes precedence (i.e. if you set different styles on the whole table and a cell, the style applied to the cell will be used).